Modified the way behavior database is created
Now, pdf files are to be added manually in ./Data/behavior/TopicFolder/raw_pdf.
I wrote some code for data scrapping from google scholar but did not include it for the github, because it looks like incompatible with google TOS and is disrupted by recaptcha and could lead to an ip ban which is annoying. Howerer, 10 articles is already around 300 pages of information about FOMO.
Run update database to update the csv
The chromadb needs to be deleted and re created which is not ideal (see rag_utils.py for the TODO), although creating the database is quite fast.
This is functional but ideally the csv should be replaced by a panda dataframe i think.